Refactor onShowMetadata to onOrderLineShowMetadata and remove barrel files
Changes:
- Renamed onShowMetadata → onOrderLineShowMetadata across 13 files for clarity
- Removed 6 barrel files (index.ts) from order components
- Updated 7 import statements to use direct component paths

The new name better describes that this handler shows metadata for order lines specifically, not for the entire order.

Barrel file removal improves code clarity by making imports explicit.
