--- src/orders/components/OrderDetailsDatagrid/index.ts
@@ -1,1 +1,0 @@
-export * from "./OrderDetailsDatagrid";
