+++ src/orders/components/OrderDetailsPage/OrderDetailsPage.tsx
@@ -33,12 +33,12 @@ import OrderChannelSectionCard from "../OrderChannelSectionCard";
 import OrderCustomer from "../OrderCustomer";
 import OrderCustomerNote from "../OrderCustomerNote";
 import OrderDraftDetails from "../OrderDraftDetails/OrderDraftDetails";
-import { FormData as OrderDraftDetailsProductsFormData } from "../OrderDraftDetailsProducts";
-import OrderFulfilledProductsCard from "../OrderFulfilledProductsCard";
+import { FormData as OrderDraftDetailsProductsFormData } from "../OrderDraftDetailsProducts/OrderDraftDetailsProducts";
+import OrderFulfilledProductsCard from "../OrderFulfilledProductsCard/OrderFulfilledProductsCard";
 import OrderHistory, { FormData as HistoryFormData } from "../OrderHistory";
 import OrderInvoiceList from "../OrderInvoiceList";
 import { OrderPaymentOrTransaction } from "../OrderPaymentOrTransaction/OrderPaymentOrTransaction";
-import OrderUnfulfilledProductsCard from "../OrderUnfulfilledProductsCard";
+import OrderUnfulfilledProductsCard from "../OrderUnfulfilledProductsCard/OrderUnfulfilledProductsCard";
 import { messages } from "./messages";
 import Title from "./Title";
 import {
@@ -65,7 +65,7 @@ interface OrderDetailsPageProps {
   onBillingAddressEdit: () => any;
   onFulfillmentApprove: (id: string) => any;
   onFulfillmentCancel: (id: string) => any;
-  onShowMetadata: (id: string) => void;
+  onOrderLineShowMetadata: (id: string) => void;
   onFulfillmentTrackingNumberUpdate: (id: string) => any;
   onOrderFulfill: () => any;
   onProductClick?: (id: string) => any;
@@ -120,7 +120,7 @@ const OrderDetailsPage = (props: OrderDetailsPageProps) => {
     onShippingMethodEdit,
     onTransactionAction,
     onAddManualTransaction,
-    onShowMetadata,
+    onOrderLineShowMetadata,
     onMarkAsPaid,
     onRefundAdd,
     onSubmit,
@@ -216,7 +216,7 @@ const OrderDetailsPage = (props: OrderDetailsPageProps) => {
                   lines={unfulfilled}
                   onFulfill={onOrderFulfill}
                   loading={loading}
-                  onShowMetadata={onShowMetadata}
+                  onOrderLineShowMetadata={onOrderLineShowMetadata}
                 />
               ) : (
                 <>
@@ -224,7 +224,7 @@ const OrderDetailsPage = (props: OrderDetailsPageProps) => {
                     order={order}
                     errors={errors}
                     loading={loading}
-                    onShowMetadata={onShowMetadata}
+                    onOrderLineShowMetadata={onOrderLineShowMetadata}
                     onOrderLineAdd={onOrderLineAdd}
                     onOrderLineChange={onOrderLineChange}
                     onOrderLineRemove={onOrderLineRemove}
@@ -240,7 +240,7 @@ const OrderDetailsPage = (props: OrderDetailsPageProps) => {
                   fulfillment={fulfillment}
                   fulfillmentAllowUnpaid={shop?.fulfillmentAllowUnpaid}
                   order={order}
-                  onShowMetadata={onShowMetadata}
+                  onOrderLineShowMetadata={onOrderLineShowMetadata}
                   onOrderFulfillmentCancel={() => onFulfillmentCancel(fulfillment.id)}
                   onTrackingCodeAdd={() => onFulfillmentTrackingNumberUpdate(fulfillment.id)}
                   onOrderFulfillmentApprove={() => onFulfillmentApprove(fulfillment.id)}
