--- src/orders/components/OrderDetailsPage/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./OrderDetailsPage";
-export * from "./OrderDetailsPage";
