+++ src/orders/components/OrderDraftDetailsProducts/OrderDraftDetailsProducts.tsx
@@ -23,7 +23,7 @@ interface OrderDraftDetailsProductsProps {
   loading: boolean;
   onOrderLineChange: (id: string, data: FormData) => void;
   onOrderLineRemove: (id: string) => void;
-  onShowMetadata: (id: string) => void;
+  onOrderLineShowMetadata: (id: string) => void;
 }
 
 const OrderDraftDetailsProducts = ({
@@ -32,7 +32,7 @@ const OrderDraftDetailsProducts = ({
   loading,
   onOrderLineChange,
   onOrderLineRemove,
-  onShowMetadata,
+  onOrderLineShowMetadata,
 }: OrderDraftDetailsProductsProps) => {
   const classes = useStyles();
   const lines = order?.lines ?? [];
@@ -49,7 +49,7 @@ const OrderDraftDetailsProducts = ({
       onOrderLineRemove={onOrderLineRemove}
       onOrderLineChange={onOrderLineChange}
       errors={formErrors}
-      onShowMetadata={onShowMetadata}
+      onOrderLineShowMetadata={onOrderLineShowMetadata}
     />
   );
 };
