--- src/orders/components/OrderDraftDetailsProducts/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./OrderDraftDetailsProducts";
-export * from "./OrderDraftDetailsProducts";
