+++ src/orders/components/OrderDraftPage/OrderDraftPage.tsx
@@ -55,7 +55,7 @@ interface OrderDraftPageProps extends FetchMoreProps {
   onShippingAddressEdit: () => void;
   onShippingMethodEdit: () => void;
   onProfileView: () => void;
-  onShowMetadata: (id: string) => void;
+  onOrderLineShowMetadata: (id: string) => void;
 }
 
 const draftOrderListUrl = orderDraftListUrl();
@@ -80,7 +80,7 @@ const OrderDraftPage = (props: OrderDraftPageProps) => {
     onShippingAddressEdit,
     onShippingMethodEdit,
     onProfileView,
-    onShowMetadata,
+    onOrderLineShowMetadata,
     order,
     channelUsabilityData,
     users,
@@ -147,7 +147,7 @@ const OrderDraftPage = (props: OrderDraftPageProps) => {
           onOrderLineChange={onOrderLineChange}
           onOrderLineRemove={onOrderLineRemove}
           onShippingMethodEdit={onShippingMethodEdit}
-          onShowMetadata={onShowMetadata}
+          onOrderLineShowMetadata={onOrderLineShowMetadata}
         />
         <OrderHistory
           history={order?.events}
