--- src/orders/components/OrderDraftPage/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./OrderDraftPage";
-export * from "./OrderDraftPage";
