--- src/orders/components/OrderFulfilledProductsCard/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./OrderFulfilledProductsCard";
-export * from "./OrderFulfilledProductsCard";
