+++ src/orders/components/OrderUnfulfilledProductsCard/OrderUnfulfilledProductsCard.tsx
@@ -6,7 +6,7 @@ import { Button, Text } from "@saleor/macaw-ui-next";
 import { FormattedMessage } from "react-intl";
 
 import OrderCardTitle from "../OrderCardTitle";
-import { OrderDetailsDatagrid } from "../OrderDetailsDatagrid";
+import { OrderDetailsDatagrid } from "../OrderDetailsDatagrid/OrderDetailsDatagrid";
 import { useStyles } from "./styles";
 import { toLineWithUnfulfilledQuantity } from "./utils";
 
@@ -16,13 +16,13 @@ interface OrderUnfulfilledProductsCardProps {
   lines: OrderLineFragment[];
   onFulfill: () => void;
   loading: boolean;
-  onShowMetadata: (id: string) => void;
+  onOrderLineShowMetadata: (id: string) => void;
 }
 
 const OrderUnfulfilledProductsCard = ({
   showFulfillmentAction,
   notAllowedToFulfillUnpaid,
-  onShowMetadata,
+  onOrderLineShowMetadata,
   lines,
   onFulfill,
   loading,
@@ -41,7 +41,7 @@ const OrderUnfulfilledProductsCard = ({
           <OrderDetailsDatagrid
             lines={toLineWithUnfulfilledQuantity(lines)}
             loading={loading}
-            onShowMetadata={onShowMetadata}
+            onOrderLineShowMetadata={onOrderLineShowMetadata}
           />
           {showFulfillmentAction && (
             <DashboardCard.BottomActions justifyContent="flex-end">
