--- src/orders/components/OrderUnfulfilledProductsCard/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./OrderUnfulfilledProductsCard";
-export * from "./OrderUnfulfilledProductsCard";
